pyproject.toml
[tool.poetry]
name = "worker-microservice"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.5"}
pyzmq = "^26.2.0"
pillow = "^11.0.0"
pydantic-settings = "^2.7.0"
sqlmodel = "^0.0.22"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
